refactor: ScatterProvider that do the dirty job alone
